+++ src/main/java/net/mohron/skyclaims/command/user/CommandCreate.java
@@ -7,7 +7,6 @@ import net.mohron.skyclaims.permissions.Options;
 import net.mohron.skyclaims.permissions.Permissions;
 import net.mohron.skyclaims.world.Island;
 import org.spongepowered.api.command.CommandException;
-import org.spongepowered.api.command.CommandPermissionException;
 import org.spongepowered.api.command.CommandResult;
 import org.spongepowered.api.command.CommandSource;
 import org.spongepowered.api.command.args.CommandContext;
@@ -16,10 +15,8 @@ import org.spongepowered.api.command.spec.CommandExecutor;
 import org.spongepowered.api.command.spec.CommandSpec;
 import org.spongepowered.api.entity.living.player.Player;
 import org.spongepowered.api.text.Text;
-import org.spongepowered.api.text.format.TextColors;
 
 public class CommandCreate implements CommandExecutor {
-
 	private static final SkyClaims PLUGIN = SkyClaims.getInstance();
 
 	public static final String HELP_TEXT = "create an island.";
@@ -29,7 +26,7 @@ public class CommandCreate implements CommandExecutor {
 	public static CommandSpec commandSpec = CommandSpec.builder()
 			.permission(Permissions.COMMAND_CREATE)
 			.description(Text.of(HELP_TEXT))
-			.arguments(GenericArguments.optional(GenericArguments.string(SCHEMATIC)))
+			.arguments(GenericArguments.optional(new SchematicArgument(SCHEMATIC)))
 			.executor(new CommandCreate())
 			.build();
 
@@ -48,28 +45,11 @@ public class CommandCreate implements CommandExecutor {
 			throw new CommandException(Text.of("You must be a player to use this command!"));
 
 		Player player = (Player) src;
-		String schematic = Options.getStringOption(player.getUniqueId(), Options.DEFAULT_SCHEMATIC);
 
 		if (Island.hasIsland(player.getUniqueId()))
 			throw new CommandException(Text.of("You already have an island!"));
 
-		if (SchematicArgument.SCHEMATICS.isEmpty())
-			throw new CommandException(Text.of("There are no valid schematics to create an island with!"));
-
-		if (args.getOne(SCHEMATIC).isPresent()) {
-			schematic = (String) args.getOne(SCHEMATIC).get();
-			if (!player.hasPermission(Permissions.COMMAND_ARGUMENTS_SCHEMATICS + "." + schematic.toLowerCase()))
-				throw new CommandPermissionException(Text.of(TextColors.RED, "You do not have permission to use the ", TextColors.YELLOW, schematic, TextColors.RED, " schematic!"));
-		}
-
-		if (!SchematicArgument.SCHEMATICS.containsKey(schematic.toLowerCase())) {
-			String schems = "";
-			for (String s : SchematicArgument.SCHEMATICS.values()) {
-				schems += s + ", ";
-			}
-			schems = schems.substring(0, schems.length() - 2);
-			throw new CommandException(Text.of("The value supplied is not a valid schematic. Choose from: ", TextColors.AQUA, schems));
-		}
+		String schematic = (String) args.getOne(SCHEMATIC).orElse(Options.getStringOption(player.getUniqueId(), Options.DEFAULT_SCHEMATIC));
 
 		player.sendMessage(Text.of("Your island is being created. You will be teleported shortly."));